requirements.txt
Click==7.0
attrs==18.2.0